% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{any_fever}
\alias{any_fever}
\title{Identify children with fever of any cause}
\usage{
any_fever(parameters, variables)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variables}
}
\value{
A bitset of children with fever
}
\description{
Identify children with fever of any cause
}
