% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{asymptomatic_event}
\alias{asymptomatic_event}
\title{Asymptomatic event}
\usage{
asymptomatic_event(variables, condition)
}
\arguments{
\item{variables}{Model variables}

\item{condition}{Condition recovered from}
}
\value{
Event
}
\description{
A move to asymptomatic. Resets status back to 1 (asymptomatic) and clears any fever.
}
