% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{background_mortality}
\alias{background_mortality}
\title{Background mortality}
\usage{
background_mortality(parameters, variables, renderer)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{renderer}{Model renderer}
}
\description{
Replace a child who dies from the background mortality rate. This is currently
a constant probability estimated using the average_age parameter.
}
