% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_chw.R
\name{chw_diagnose_diarrhoea}
\alias{chw_diagnose_diarrhoea}
\title{CHW diarrhoea diagnosis}
\usage{
chw_diagnose_diarrhoea(target, variables, parameters)
}
\arguments{
\item{target}{Children presenting at the CHW}

\item{variables}{Model variables}

\item{parameters}{Model parameters}
}
\value{
A bit set of children that have tested +ve (including false +ves) and
who have been successfully processed by the health facility.
}
\description{
Accounts for the diagnostic a CHW uses as well as general human error
}
