% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_chw.R
\name{chw_diagnose_severe_pneumonia}
\alias{chw_diagnose_severe_pneumonia}
\title{CHW facility severe pneumonia diagnosis}
\usage{
chw_diagnose_severe_pneumonia(target, variables, parameters, timestep)
}
\arguments{
\item{target}{Children presenting at the CHW}

\item{variables}{Model variables}

\item{parameters}{Model parameters}

\item{timestep}{Timestep}
}
\value{
A bit set of children that have tested +ve (including false +ves) and
who have been successfully processed by the health facility.
}
\description{
Accounts for the diagnostic a health facility uses and long-symptom diagnosis
as well as general human error
}
