% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_chw.R
\name{chw_treat}
\alias{chw_treat}
\title{Community Health Worker treatment}
\usage{
chw_treat(variables, parameters, renderer, events)
}
\arguments{
\item{variables}{Model variables}

\item{parameters}{Model parameters}

\item{renderer}{Model renderer}

\item{events}{Model events}
}
\description{
Treats children coming to a CHW. Severe cases take precedent and are referred if detected, other
outcomes depend on protocol, diagnostics and treatments.
}
