% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{clear_all_scheduled_events}
\alias{clear_all_scheduled_events}
\title{Clear all scheduled events for a given target}
\usage{
clear_all_scheduled_events(target, events)
}
\arguments{
\item{target}{Children to clear events for}

\item{events}{Model events}
}
\description{
Clear all scheduled events for a given target
}
