% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{clear_scheduled_treatment_visits}
\alias{clear_scheduled_treatment_visits}
\title{Clear any scheduled treatment visits}
\usage{
clear_scheduled_treatment_visits(target, events)
}
\arguments{
\item{target}{Children}

\item{events}{Model events}
}
\description{
Clear any scheduled treatment visits
}
