% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity.R
\name{clinical_immunity}
\alias{clinical_immunity}
\title{Clinical immunity}
\usage{
clinical_immunity(exposure, shape, rate)
}
\arguments{
\item{exposure}{Number of previous infections}

\item{shape}{Shape parameter}

\item{rate}{Rate parameter (for a fixed shape, a higher rate will lead to a faster build up of immunity)}
}
\value{
Exposure immunity modifier
}
\description{
Modifier of clinical disease 0 (all asymptomatic) and 1 (all symptomatic), dependent on
a child's prior exposure.
}
