% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{community_impact}
\alias{community_impact}
\title{Intervention community impact modifier}
\usage{
community_impact(type, index, p)
}
\arguments{
\item{type}{Infection type}

\item{index}{Infection type index}

\item{p}{Condition specific parameters}
}
\description{
Intervention community impact modifier
}
