% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{condition_exposure}
\alias{condition_exposure}
\title{Condition exposure}
\usage{
condition_exposure(condition, individuals, variables, parameters, events)
}
\arguments{
\item{condition}{Condition: diarrhoea, malaria or pneumonia}

\item{individuals}{Model individuals}

\item{variables}{Model variables}

\item{parameters}{Model parameters}

\item{events}{Model events}
}
\description{
Implements exposure to one of three conditions (diarrhoea, malaria, pneumonia). Selects
those infected, selects which disease they have been infected with and schedules the disease
life course.
}
