% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{convert_to_long}
\alias{convert_to_long}
\title{Convert wide output to long.}
\usage{
convert_to_long(output)
}
\arguments{
\item{output}{Simulation output from \code{individual::simulate}}
}
\value{
Long output
}
\description{
Convert wide output to long.
}
