% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{create_disease_variables}
\alias{create_disease_variables}
\title{Create disease variables}
\usage{
create_disease_variables(parameters)
}
\arguments{
\item{parameters}{Model parameters}
}
\value{
A list of disease variables
}
\description{
Creates all variables associated with disease.
}
