% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{create_events}
\alias{create_events}
\title{Create events}
\usage{
create_events(individuals, variables)
}
\arguments{
\item{individuals}{Model individuals}

\item{variables}{Model variables}
}
\value{
List of events
}
\description{
Create events
}
