% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{create_individuals}
\alias{create_individuals}
\title{Create individuals}
\usage{
create_individuals(states, variables, events)
}
\arguments{
\item{states}{Model states}

\item{variables}{Model variables}

\item{events}{Model events}
}
\value{
List of individuals
}
\description{
Create individuals
}
