% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{create_processes}
\alias{create_processes}
\title{Create list of model processes}
\usage{
create_processes(parameters, variables, renderer, events)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{renderer}{Model renderer}

\item{events}{Model events}
}
\value{
List of model process functions
}
\description{
Create list of model processes
}
