% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{create_variables}
\alias{create_variables}
\title{Create individual variables}
\usage{
create_variables(parameters)
}
\arguments{
\item{parameters}{Model parameters}
}
\description{
Create individual variables
}
