% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{cure}
\alias{cure}
\title{Cure a condition}
\usage{
cure(target, disease, variables, events)
}
\arguments{
\item{target}{Target children}

\item{disease}{Disease index}

\item{variables}{Model variables}

\item{events}{Model events}
}
\description{
Cure a condition
}
