% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{death_from_diseases}
\alias{death_from_diseases}
\title{Death from diseases}
\usage{
death_from_diseases(
  death_index,
  parameters,
  variables,
  events,
  renderer,
  timestep
)
}
\arguments{
\item{death_index}{Index of cause of deaths}

\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{events}{Model events}

\item{renderer}{Model renderer}

\item{timestep}{Model timestep}
}
\description{
Death from diseases
}
