% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{death_from_other_causes}
\alias{death_from_other_causes}
\title{Death from other (non-modelled disease) causes}
\usage{
death_from_other_causes(
  death_index,
  parameters,
  variables,
  events,
  renderer,
  timestep
)
}
\arguments{
\item{death_index}{Index of cause of deaths}

\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{events}{Model events}

\item{renderer}{Model renderer}

\item{timestep}{Model timestep}
}
\description{
Death from other (non-modelled disease) causes
}
