% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{diagnosis}
\alias{diagnosis}
\title{Diagnostic result for disease}
\usage{
diagnosis(
  target,
  sens,
  spec,
  parameters,
  variables,
  disease_index,
  positive_index = c("asymptomatic", "symptomatic", "severe")
)
}
\arguments{
\item{target}{Children visiting provider}

\item{sens}{Diagnostic sensitivity}

\item{spec}{Diagnostic specificity}

\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{disease_index}{Disease index}

\item{positive_index}{State that are positive}
}
\value{
A bitset of those testing positive
}
\description{
Diagnostic result for disease
}
