% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{die}
\alias{die}
\title{Death from disease}
\usage{
die(condition, parameters, variables, events, renderer)
}
\arguments{
\item{condition}{Condition: diarrhoea, malaria or pneumonia}

\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{events}{Model events}

\item{renderer}{Model renderer}
}
\description{
Sample severely-infected individuals to die
}
