% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{dx}
\alias{dx}
\title{Diagnostic result}
\usage{
dx(status, sens, spec, positive = 1:3)
}
\arguments{
\item{status}{Child status}

\item{sens}{Diagnostic sensitivity}

\item{spec}{Diagnostic specificity}

\item{positive}{Status categories associated that are true +ves}
}
\value{
Boolean vector, TRUE = positive test, FALSE = negative test
}
\description{
Returns a boolean indicating results of a diagnostic test
}
