% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eq_prior_indiv}
\alias{eq_prior_indiv}
\title{Childs equilibrium number of previous infections}
\usage{
eq_prior_indiv(age, sigma, het, vx, mi, ii_shape, ii_rate)
}
\arguments{
\item{age}{Age}

\item{sigma}{FOI}

\item{het}{Heterogeneity}

\item{vx}{vaccine impact wrt age (pre compute across whole age range)}

\item{mi}{Maternal immunity impact (pre computre across whole age range)}

\item{ii_shape}{Infection immunity shape parameter}

\item{ii_rate}{Infection immunity rate parameter}
}
\description{
Childs equilibrium number of previous infections
}
