% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{exposure}
\alias{exposure}
\title{Exposure process}
\usage{
exposure(variables, parameters, events, renderer)
}
\arguments{
\item{variables}{Model variables}

\item{parameters}{Model parameters}

\item{events}{Model events}

\item{renderer}{Model renderer}
}
\description{
Governs exposure of all diseases.
}
