% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity.R
\name{exposure_immunity}
\alias{exposure_immunity}
\title{Exposure-driven immunity}
\usage{
exposure_immunity(exposure, shape, rate)
}
\arguments{
\item{exposure}{Number of previous infections}

\item{shape}{Shape parameter}

\item{rate}{Rate parameter (for a fixed shape, a higher rate will lead to a faster build up of immunity)}
}
\value{
Exposure immunity modifier
}
\description{
Modifier of infection 0 (all blocked) and 1 (none-blocked), dependent on
a child's prior exposure.
}
