% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{get_age}
\alias{get_age}
\title{Get children's ages}
\usage{
get_age(api, individuals, variables, index = NULL)
}
\arguments{
\item{api}{Model API}

\item{individuals}{Model individuals}

\item{variables}{Model variables}

\item{index}{optionally return a subset of the variable vector}
}
\description{
Get children's ages
}
