% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get model parameters}
\usage{
get_parameters(overrides = list())
}
\arguments{
\item{overrides}{a named list of parameter values to use instead of defaults.}
}
\description{
Create a named list of parameters for use in the model.

The parameters are defined below.

\strong{Demography:}
\itemize{
\item pop - Population size
\item lifespan - The average lifespan of an individual within the population
}
}
