% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{give_act}
\alias{give_act}
\title{Give ACT}
\usage{
give_act(target, parameters, variables, events, timestep)
}
\arguments{
\item{target}{Target children}

\item{parameters}{Model parameters}

\item{variables}{Model variable}

\item{events}{Model events}

\item{timestep}{Model timestep}
}
\description{
Provides artemisinin combination therapy to child
}
