% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{give_amoxicillin}
\alias{give_amoxicillin}
\title{Give amoxicillin}
\usage{
give_amoxicillin(target, parameters, variables, events, timestep)
}
\arguments{
\item{target}{Target children}

\item{parameters}{Model parameters}

\item{variables}{Model variable}

\item{events}{Model events}

\item{timestep}{Model timestep}
}
\description{
Provides amoxicillin antibiotic therapy to child
}
