% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{give_ors}
\alias{give_ors}
\title{Give ORS}
\usage{
give_ors(target, parameters, variables, events, timestep)
}
\arguments{
\item{target}{Target children}

\item{parameters}{Model parameters}

\item{variables}{Model variable}

\item{events}{Model events}

\item{timestep}{Model timestep}
}
\description{
Provides oral rehydration salts (ORS) to a child
}
