% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{give_severe_treatment_malaria}
\alias{give_severe_treatment_malaria}
\title{Give treatment for severe malaria}
\usage{
give_severe_treatment_malaria(target, parameters, variables, events, timestep)
}
\arguments{
\item{target}{Target children}

\item{parameters}{Model parameters}

\item{variables}{Model variable}

\item{events}{Model events}

\item{timestep}{Model timestep}
}
\description{
Provides treatment for severe malaria
}
