% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{give_severe_treatment_pneumonia}
\alias{give_severe_treatment_pneumonia}
\title{Give treatment for severe pneumonia}
\usage{
give_severe_treatment_pneumonia(
  target,
  parameters,
  variables,
  events,
  timestep
)
}
\description{
Provides treatment for severe pneumonia
}
