% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{graduate_event}
\alias{graduate_event}
\title{Graduate event}
\usage{
graduate_event(variables, parameters, events, renderer)
}
\arguments{
\item{variables}{Model variables}

\item{parameters}{Model parameters}

\item{events}{Model events}

\item{renderer}{Model renderer}
}
\value{
Event
}
\description{
Replaces a child that has reached age five with a newborn
}
