% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity.R
\name{heterogeneity}
\alias{heterogeneity}
\title{Draw individual level heterogeneities}
\usage{
heterogeneity(n, het_sd)
}
\arguments{
\item{n}{Number}

\item{het_sd}{Standard deviation}
}
\description{
Draw individual level heterogeneities
}
