% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_hf.R
\name{hf_diagnose_pneumonia}
\alias{hf_diagnose_pneumonia}
\title{Health facility pneumonia diagnosis}
\usage{
hf_diagnose_pneumonia(target, variables, parameters)
}
\arguments{
\item{target}{Children presenting at the health facility}

\item{variables}{Model variables}

\item{parameters}{Model parameters}
}
\value{
A bit set of children that have tested +ve (including false +ves) and
who have been successfully processed by the health facility.
}
\description{
Accounts for the diagnostic a health facility uses as well as general human error
}
