% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_hf.R
\name{hf_diagnose_severe_diarrhoea}
\alias{hf_diagnose_severe_diarrhoea}
\title{Health facility severe diarrhoea diagnosis}
\usage{
hf_diagnose_severe_diarrhoea(target, variables, parameters)
}
\arguments{
\item{target}{Children presenting at the health facility}

\item{variables}{Model variables}

\item{parameters}{Model parametyers}
}
\value{
A bit set of children that have tested +ve (including false +ves) and
who have been successfully processed by the health facility.
}
\description{
Accounts for the diagnostic a health facility uses as well as general human error
}
