% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_hf.R
\name{hf_treat}
\alias{hf_treat}
\title{Health Facility treatment}
\usage{
hf_treat(variables, parameters, renderer, events)
}
\arguments{
\item{variables}{Model variables}

\item{parameters}{Model parameters}

\item{renderer}{Model renderer}

\item{events}{Model events}
}
\description{
Treats children coming to the health facility. Severe cases take precedent, other
outcomes depend on diagnostics and treatments.
}
