% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_hf.R
\name{hf_treat_severe_diarrhoea}
\alias{hf_treat_severe_diarrhoea}
\title{Health facility treatment for severe diarrhoea}
\usage{
hf_treat_severe_diarrhoea(target, variables, parameters, events)
}
\arguments{
\item{target}{Children presenting at the health facility}

\item{variables}{Model variables}

\item{parameters}{Model parameters}

\item{events}{Model events}
}
\description{
Health facility treatment for severe diarrhoea
}
