% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_hf.R
\name{hf_treat_severe_malaria}
\alias{hf_treat_severe_malaria}
\title{Health facility treatment for severe malaria}
\usage{
hf_treat_severe_malaria(target, variables, parameters, events)
}
\description{
Health facility treatment for severe malaria
}
