% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_hf.R
\name{hf_treat_severe_pneumonia}
\alias{hf_treat_severe_pneumonia}
\title{Health facility treatment for severe pneumonia}
\usage{
hf_treat_severe_pneumonia(target, variables, parameters, events)
}
\description{
Health facility treatment for severe pneumonia
}
