% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{increment_counter}
\alias{increment_counter}
\title{Increment an integer variable +1}
\usage{
increment_counter(target, variable)
}
\arguments{
\item{target}{Bitset of individuals}

\item{variable}{Variable to increment}
}
\description{
Increment an integer variable +1
}
