% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{increment_prior_exposure_counter}
\alias{increment_prior_exposure_counter}
\title{Record new infections in individual's history of infection.}
\usage{
increment_prior_exposure_counter(
  new_infections,
  target,
  prior_labels,
  variables
)
}
\arguments{
\item{new_infections}{Indices of new infections}

\item{target}{Bitset of newly infected individuals}

\item{prior_labels}{Names of prior variables}

\item{variables}{Model variables}
}
\description{
Record new infections in individual's history of infection.
}
