% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{infection}
\alias{infection}
\title{Infection}
\usage{
infection(infected, disease, parameters, variables, events, renderer, timestep)
}
\arguments{
\item{infected}{Target bitset of those infected}

\item{disease}{Disease}

\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{events}{Model events}

\item{renderer}{Model renderer}

\item{timestep}{Model timestep}
}
\description{
Infection of children (to asymptomatic or symptomatic states).
}
