% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{infection_life_course}
\alias{infection_life_course}
\title{Schedule the disease life course}
\usage{
infection_life_course(
  condition,
  infection_disease,
  p,
  target,
  api,
  individuals,
  variables,
  events
)
}
\arguments{
\item{condition}{Condition: diarrhoea, malaria or pneumonia}

\item{infection_disease}{Inidces of disease type}

\item{p}{Condition-specific subset of parameters}

\item{target}{Indices of children}

\item{api}{Model API}

\item{individuals}{Model individuals}

\item{variables}{Model variables}

\item{events}{Model events}
}
\description{
Schedule the disease life course
}
