% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{infection_probability}
\alias{infection_probability}
\title{Infection probability vector}
\usage{
infection_probability(target, disease, parameters, variables, timestep)
}
\arguments{
\item{target}{Those exposed}

\item{timestep}{Timestep}

\item{p}{Disease parameters}

\item{prior_exposure}{Prior exposure to disease}

\item{heterogeneity}{Individual heterogeneity}

\item{birth_t}{Birth times}
}
\description{
Infection probability vector
}
