% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{initialise_events}
\alias{initialise_events}
\title{Initialise events}
\usage{
initialise_events(events, variables, parameters)
}
\arguments{
\item{events}{Model events}

\item{variables}{Model variables}

\item{parameters}{Model parameters}
}
\description{
Run any scheduling that is required when initialising the population
}
