% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{long_symptoms}
\alias{long_symptoms}
\title{Long symptoms}
\usage{
long_symptoms(symptom_start_var, target, threshold, timestep)
}
\arguments{
\item{symptom_start_var}{Variable of symptom start time}

\item{target}{Children seeking treatment}

\item{threshold}{Period after which illness is classified as severe}

\item{timestep}{Timestyep}
}
\value{
Bitset
}
\description{
Returns a bitset of those who have had symptoms for a period > the threshold to define illness as severe
}
