% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity.R
\name{maternal_immunity}
\alias{maternal_immunity}
\title{Maternal immunity}
\usage{
maternal_immunity(age, hl)
}
\arguments{
\item{age}{Age}

\item{hl}{Halflife of duration of maternal immunity}
}
\value{
Maternal immunity modifier
}
\description{
Modifier of infection between 0 (all blocked) and 1 (none-blocked), dependent on
a child's age.
}
