% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{mortality}
\alias{mortality}
\title{Mortality}
\usage{
mortality(parameters, variables, events, renderer)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{events}{Model events}

\item{renderer}{Model renderer}
}
\description{
Samples any death events, assigns a cause and replaces children who have died.
}
