% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{mortality_probability}
\alias{mortality_probability}
\title{Estimate matrix of mortality probabilities}
\usage{
mortality_probability(parameters, variables)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variables}
}
\description{
Estimate matrix of mortality probabilities
}
