% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{nested_list_names}
\alias{nested_list_names}
\title{Extract names from nested list}
\usage{
nested_list_names(x)
}
\arguments{
\item{x}{List}
}
\value{
Names from nested list. Levels are separated by ".".
}
\description{
Extract names from nested list
}
