% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{overwrite_params}
\alias{overwrite_params}
\title{Overwrite default parameters}
\usage{
overwrite_params(parameters, user_overwrite)
}
\arguments{
\item{parameters}{Default parameters}

\item{user_overwrite}{User input (nested) list of parameters}
}
\value{
Set of user-modified simulation parameters
}
\description{
Overwrite default parameters
}
