% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sim}
\alias{plot_sim}
\title{Plot simulation output}
\usage{
plot_sim(output, return_list = FALSE)
}
\arguments{
\item{output}{Long simulation output}

\item{return_list}{Return list of component plots}
}
\value{
Output plot, one panel for each variable
}
\description{
Plot simulation output
}
