% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{prior_exposure_equilibrium}
\alias{prior_exposure_equilibrium}
\title{Prior exposure equilibrium}
\usage{
prior_exposure_equilibrium(p, population, initial_age, maximum_age, est_het)
}
\arguments{
\item{p}{Condition parameters}

\item{initial_age}{Age of each child}

\item{maximum_age}{Maximum age}

\item{est_het}{Heterogeneity of each child variable}
}
\description{
Prior exposure equilibrium
}
