% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{prior_exposure_matrix}
\alias{prior_exposure_matrix}
\title{Prior exposure equilibrium matrix}
\usage{
prior_exposure_matrix(p, size, initial_age, est_het)
}
\arguments{
\item{p}{Condition parameters}

\item{size}{Population size}

\item{initial_age}{Age variable}

\item{est_het}{Het variable}
}
\description{
Prior exposure equilibrium matrix
}
