% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_private.R
\name{private_diagnose_malaria}
\alias{private_diagnose_malaria}
\title{Private provider malaria diagnosis}
\usage{
private_diagnose_malaria(target, variables, parameters)
}
\arguments{
\item{target}{Children presenting at the Private}

\item{variables}{Model variables}

\item{parameters}{Model parameters}
}
\value{
A bit set of children that have tested +ve (including false +ves) and
who have been successfully processed by the health facility.
}
\description{
Accounts for the diagnostic a private provider uses as well as general human error
}
