% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_private.R
\name{private_treat}
\alias{private_treat}
\title{Private treatment}
\usage{
private_treat(variables, parameters, renderer, events)
}
\arguments{
\item{variables}{Model variables}

\item{parameters}{Model parameters}

\item{renderer}{Model renderer}

\item{events}{Model events}
}
\description{
Treats children coming to a private provider.
}
