% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_to_rate}
\alias{prob_to_rate}
\title{Convert probability to rate}
\usage{
prob_to_rate(p, t = 1)
}
\arguments{
\item{p}{probability}

\item{t}{time period}
}
\value{
Probability of event in timestep
}
\description{
The constant rate at which an event with a given probability will occur
}
