% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{progress_severe}
\alias{progress_severe}
\title{Progress to severe disease}
\usage{
progress_severe(condition, parameters, variables, events)
}
\arguments{
\item{condition}{Condition: diarrhoea, malaria or pneumonia}

\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{events}{Model events}
}
\description{
Sample infected individuals to progress to severe disease
}
