% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider.R
\name{px}
\alias{px}
\title{Provider result}
\usage{
px(disease, efficacy)
}
\arguments{
\item{disease}{Disease}

\item{efficacy}{Provider efficacy}
}
\description{
A function to capture the efficacy of the provider - currently encapsulates all aspects of "human error".
}
