% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{rate_to_prob}
\alias{rate_to_prob}
\title{Convert rate to probability}
\usage{
rate_to_prob(r, t = 1)
}
\arguments{
\item{r}{rate}

\item{t}{time period}
}
\value{
Probability of event in timestep
}
\description{
The probability a event with a constant rate will occur within a given time period
}
