% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{render_demography}
\alias{render_demography}
\title{Render demographic outputs}
\usage{
render_demography(individuals, variables)
}
\arguments{
\item{individuals}{Model individuals}

\item{variables}{Model variables}
}
\description{
Average ages and number of children in each year age-group.
}
