% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_fever_prevalence}
\alias{render_fever_prevalence}
\title{Record prevalence of fever}
\usage{
render_fever_prevalence(parameters, variables, renderer)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{renderer}{Model renderer}
}
\description{
Record prevalence of fever
}
