% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{render_fevers}
\alias{render_fevers}
\title{Record fever prevalence}
\usage{
render_fevers(variables, parameters, renderer)
}
\arguments{
\item{variables}{Model variables}

\item{parameters}{Model parameters}

\item{renderer}{Model renderer}
}
\description{
Record prevalence of fever
}
