% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{render_prevalence}
\alias{render_prevalence}
\title{Render prevalence outputs for a condition}
\usage{
render_prevalence(condition, variables, parameters, renderer)
}
\arguments{
\item{condition}{Condition: diarrhoea, malaria or pneumonia}

\item{variables}{Model variables}

\item{parameters}{Model parameters}

\item{renderer}{Model renderer}
}
\description{
Render prevalence outputs for a condition
}
