% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_prior}
\alias{render_prior}
\title{Record average prior infections}
\usage{
render_prior(variables, renderer, parameters)
}
\arguments{
\item{variables}{Model variables}

\item{renderer}{Model renderer}

\item{parameters}{Model parameters}
}
\description{
Record average prior infections
}
