% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{render_prior_exposure}
\alias{render_prior_exposure}
\title{Record prior exposure}
\usage{
render_prior_exposure(condition, variables, parameters, renderer)
}
\arguments{
\item{condition}{Condition: diarrhoea, malaria or pneumonia}

\item{variables}{Model variables}

\item{parameters}{Model parameters}

\item{renderer}{Model renderer}
}
\description{
Record the average number of prior disease-specific infections
}
