% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{replace_child}
\alias{replace_child}
\title{Replace child with new}
\usage{
replace_child(target, timestep, variables, parameters, events)
}
\arguments{
\item{target}{Target indices}

\item{timestep}{Current time}

\item{variables}{Model variables}

\item{parameters}{Model parameters}

\item{events}{Model events}
}
\description{
Replace child with new
}
