% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{replace_render_na}
\alias{replace_render_na}
\title{Replace NA in output}
\usage{
replace_render_na(
  output,
  variables = c("chw_patients", "chw_ors", "chw_followup", "private_patients",
    "private_ors", "hf_patients", "hf_ors", "hf_severe_diarrhoea_tx", "graduation")
)
}
\arguments{
\item{output}{Model output}

\item{variables}{Variables where we want to convert NAs -> 0s}
}
\description{
When rendering occurs in a function that is not called every timestep, missing outputs
are NA. We can selectivly convert these to zeros
}
