% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{resample_preference}
\alias{resample_preference}
\title{Resample provider preference when CHW introduced}
\usage{
resample_preference(parameters, variables)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variables}
}
\description{
Resample provider preference when CHW introduced
}
