% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{reset_demography}
\alias{reset_demography}
\title{Reset child demographic variables}
\usage{
reset_demography(target, parameters, variables, events, timestep)
}
\arguments{
\item{target}{Target children}

\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{events}{Model events}

\item{timestep}{Current timestep}
}
\description{
Reset child demographic variables
}
