% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{reset_disease_status}
\alias{reset_disease_status}
\title{Reset disease status for a new child}
\usage{
reset_disease_status(target, parameters, variables)
}
\arguments{
\item{target}{Target children}

\item{parameters}{Model parameters}

\item{variables}{Model variables}
}
\description{
Reset disease status for a new child
}
