% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{reset_heterogeneity}
\alias{reset_heterogeneity}
\title{Reset child heterogeneity}
\usage{
reset_heterogeneity(target, parameters, variables)
}
\arguments{
\item{target}{Target children}

\item{parameters}{Model parameters}

\item{variables}{Model variables}
}
\description{
Reset child heterogeneity
}
