% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{reset_treatment_and_interventions}
\alias{reset_treatment_and_interventions}
\title{Reset treatment and interventions}
\usage{
reset_treatment_and_interventions(target, parameters, variables)
}
\arguments{
\item{target}{Target children}

\item{parameters}{Model parameters}

\item{variables}{Model variables}
}
\description{
Reset treatment and interventions
}
