% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{rtexp}
\alias{rtexp}
\title{Random deviates from upper-truncated exponential distribution}
\usage{
rtexp(n = 1, rate = 1, lower = 0, upper = Inf)
}
\arguments{
\item{n}{Number of draws}

\item{rate}{Rate}

\item{lower}{Lower truncation (inclusive: \link{lower, upper})}

\item{upper}{Upper truncation (inclusive: \link{lower, upper})}
}
\value{
Random draws from truncated exponential
}
\description{
Random deviates from upper-truncated exponential distribution
}
