% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_simulation}
\alias{run_simulation}
\title{Run an iccm simulation}
\usage{
run_simulation(timesteps, parameters = NULL, long = TRUE)
}
\arguments{
\item{timesteps}{The number of timesteps (days) to run the model.}

\item{parameters}{Model parameter default overrides}

\item{long}{Return long output?}
}
\value{
Simulation output
}
\description{
Run an iccm simulation
}
