% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{sample_disease}
\alias{sample_disease}
\title{Sample disease if infected with a condition}
\usage{
sample_disease(p, n)
}
\arguments{
\item{p}{Disease-specific infection probabilities}

\item{n}{Number of diseases}
}
\value{
Sampled disease that will infect the individual
}
\description{
Sample disease if infected with a condition
}
