% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{sample_preference}
\alias{sample_preference}
\title{Sample provider preference}
\usage{
sample_preference(n, parameters)
}
\arguments{
\item{n}{Number of children}

\item{parameters}{Model parameters}
}
\value{
Vector of provider preference for each child
}
\description{
Used to select the provider preference assigned to each child at birth. Provider
choice can be weighted. Providers are currently: HF, CHW or Private. If no providers
are available preference is set to "None".
}
