% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{severe_diagnosis}
\alias{severe_diagnosis}
\title{Diagnostic result for severe disease}
\usage{
severe_diagnosis(target, sens, spec, parameters, variables, disease_index)
}
\arguments{
\item{target}{Children visiting provider}

\item{sens}{Diagnostic sensitivity}

\item{spec}{Diagnostic specificity}

\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{disease_index}{Disease index}
}
\value{
A bitset of those testing positive
}
\description{
Diagnostic result for severe disease
}
