% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{severity}
\alias{severity}
\title{Severity process}
\usage{
severity(variables, parameters, renderer)
}
\arguments{
\item{variables}{Model variables}

\item{parameters}{Model parameters}

\item{renderer}{Model renderer}
}
\description{
Severity process
}
