% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sub_plot}
\alias{sub_plot}
\title{Sub plot}
\usage{
sub_plot(output, group_name, group_vars, var_names, ylab, smooth = TRUE)
}
\arguments{
\item{output}{Model output (long format)}

\item{group_name}{Name of group}

\item{group_vars}{Names of variables to plot together}

\item{var_names}{Vector of variable names to relabel variables with}

\item{ylab}{Y axis label}

\item{smooth}{Apply 2 month rolling average smoothing}
}
\value{
A sub plot
}
\description{
Create sub-plots from names variables
}
