% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{treatment_prophylaxis}
\alias{treatment_prophylaxis}
\title{Treatment prophylaxis modifier}
\usage{
treatment_prophylaxis(target, disease, parameters, variables, timestep)
}
\arguments{
\item{target}{Children exposed}

\item{disease}{Disease index}

\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{timestep}{Current timestep}
}
\value{
Treatment prophylaxis effect
}
\description{
Treatment prophylaxis modifier
}
