% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{vaccine_effect}
\alias{vaccine_effect}
\title{Vaccine effect}
\usage{
vaccine_effect(ages, vx_start, vx_initial_efficacy, vx_hl)
}
\arguments{
\item{ages}{Ages of children}

\item{vx_start}{Age of vaccinated child when fully protected}

\item{vx_initial_efficacy}{Initial (maximum) efficacy}

\item{vx_hl}{Vaccine protection half life}
}
\value{
Vaccine effects
}
\description{
Estimates the level of protection from vaccination given time since vaccination and vaccine properties.
}
