% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{any_fever}
\alias{any_fever}
\title{Any fever}
\usage{
any_fever(variables)
}
\arguments{
\item{variables}{Model variables}
}
\value{
Bitset
}
\description{
Returns a bitset of individuals who have fever of any cause
}
