% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{community_impact}
\alias{community_impact}
\title{Intervention community impact modifier}
\usage{
community_impact(disease, index, p)
}
\arguments{
\item{disease}{Infection disease}

\item{index}{Infection disease index}

\item{p}{Condition specific parameters}
}
\description{
Intervention community impact modifier
}
