% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{create_event_listeners}
\alias{create_event_listeners}
\title{Adds listeners to specific events}
\usage{
create_event_listeners(events, variables, parameters, renderer)
}
\arguments{
\item{events}{Model events}

\item{variables}{Model variables}

\item{parameters}{Model parameters}

\item{renderer}{Model renderer}
}
\description{
Adds listeners to specific events
}
