% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{create_events}
\alias{create_events}
\title{Create events}
\usage{
create_events(variables, parameters)
}
\arguments{
\item{variables}{Model variables}

\item{parameters}{Model parameters}
}
\value{
List of events
}
\description{
Create events
}
