% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{create_processes}
\alias{create_processes}
\title{Create list of model processes}
\usage{
create_processes(parameters, individuals, variables)
}
\arguments{
\item{parameters}{Model parameters}

\item{individuals}{Model individuals}

\item{variables}{Model variables}
}
\value{
List of model process functions
}
\description{
Create list of model processes
}
