% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{cure}
\alias{cure}
\title{Cure a condition}
\usage{
cure(target, condition, variables, events)
}
\arguments{
\item{target}{Target children}

\item{condition}{Condition to cure}

\item{variables}{Model variable}

\item{events}{Model events}
}
\description{
Cure a condition
}
