% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{diagnosis}
\alias{diagnosis}
\title{Diagnostic result for disease}
\usage{
diagnosis(target, sens, spec, parameters, variables, disease_index)
}
\arguments{
\item{target}{Children visiting provider}

\item{sens}{Diagnostic sensitivity}

\item{spec}{Diagnostic specificity}

\item{disease_index}{Index of diseases of target type}

\item{infection_status}{All infection status variables}
}
\value{
A bitset of those testing positive
}
\description{
Diagnostic result for disease
}
