% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{dx}
\alias{dx}
\title{Diagnostic result}
\usage{
dx(target, status_variable, sens, spec, positive, negative)
}
\arguments{
\item{target}{Children visiting provider}

\item{status_variable}{Disease status variable}

\item{sens}{Diagnostic sensitivity}

\item{spec}{Diagnostic specificity}

\item{positive}{Status categories that are true +ves}

\item{negative}{Status categories  that are true -ves}
}
\value{
Bitset
}
\description{
Returns a bitset of those testing positive
}
