% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{give_severe_treatment_diarrhoea}
\alias{give_severe_treatment_diarrhoea}
\title{Give treatment for severe diarrhoea}
\usage{
give_severe_treatment_diarrhoea(target, efficacy)
}
\arguments{
\item{target}{Target children}
}
\description{
Provides treatment for severe diarrhoea
}
