% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{give_severe_treatment_malaria}
\alias{give_severe_treatment_malaria}
\title{Give treatment for severe malaria}
\usage{
give_severe_treatment_malaria(target, efficacy)
}
\arguments{
\item{target}{Target children}
}
\description{
Provides treatment for severe malaria
}
