% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{graduate}
\alias{graduate}
\title{Graduation}
\usage{
graduate(parameters, individuals, variables)
}
\arguments{
\item{parameters}{Model parameters}

\item{individuals}{Model individuals}

\item{variables}{Model variables}
}
\description{
Replace a child who reaches the maximum age in the simulation with a new child.
}
