% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{graduate}
\alias{graduate}
\title{Graduation}
\usage{
graduate(parameters, variables, renderer)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{renderer}{Model renderer}
}
\description{
Replace a child who reaches the maximum age in the simulation with a new child.
}
