% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{infection_life_course}
\alias{infection_life_course}
\title{Schedule the disease life course}
\usage{
infection_life_course(
  condition,
  infection_type_index,
  priors,
  p,
  target,
  variables,
  events
)
}
\arguments{
\item{condition}{Condition: diarrhoea, malaria or pneumonia}

\item{infection_type_index}{Indices of disease type}

\item{priors}{Name of prior exposure variables}

\item{p}{Condition-specific subset of parameters}

\item{target}{Indices of children}

\item{variables}{Model variables}

\item{events}{Model events}
}
\description{
Schedule the disease life course
}
