% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{infection_life_course}
\alias{infection_life_course}
\title{Schedule the disease life course}
\usage{
infection_life_course(
  condition,
  type,
  target,
  prior_name,
  duration,
  variables,
  events,
  renderer,
  timestep
)
}
\arguments{
\item{condition}{Condition: diarrhoea, malaria or pneumonia}

\item{type}{Infection type}

\item{target}{Target children}

\item{prior_name}{Name of prior variable}

\item{duration}{Durations of clinical episodes}

\item{variables}{Model variables}

\item{events}{Model events}

\item{renderer}{Model renderer}

\item{timestep}{Timestep}
}
\description{
Schedule the disease life course
}
