% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{infection_probability}
\alias{infection_probability}
\title{Infection probability vector}
\usage{
infection_probability(target, disease, parameters, variables, timestep)
}
\arguments{
\item{target}{Those exposed}

\item{disease}{disease index}

\item{parameters}{Model parameters}

\item{variables}{Model variables}

\item{timestep}{Timestep}
}
\description{
Infection probability vector
}
