% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{infection_probability}
\alias{infection_probability}
\title{Infection probability vector}
\usage{
infection_probability(disease, target, p, timestep, variables)
}
\arguments{
\item{disease}{Disease}

\item{target}{Those exposed}

\item{p}{Disease parameters}

\item{timestep}{Model timestep}

\item{variables}{Model variables}
}
\description{
Infection probability vector
}
