% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{initialise_render_defaults}
\alias{initialise_render_defaults}
\title{Add default values to render variables that won't get called every timestep}
\usage{
initialise_render_defaults(renderer, parameters)
}
\arguments{
\item{renderer}{Model renderer}
}
\description{
Add default values to render variables that won't get called every timestep
}
