% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{initialise_render_defaults}
\alias{initialise_render_defaults}
\title{Add default values to render variables that won't get called every timestep}
\usage{
initialise_render_defaults(
  renderer,
  zero_default = c("chw_patients", "chw_ors", "chw_followup", "chw_referral",
    "private_patients", "private_ors", "hf_patients", "hf_ors", "hf_severe_diarrhoea_tx",
    "graduation", "dia_bacteria_mortality", "dia_virus_mortality",
    "dia_parasite_mortality", "dia_rotavirus_mortality", "malaria_pf_mortality")
)
}
\arguments{
\item{renderer}{Model renderer}

\item{zero_default}{Variables to set a default 0}
}
\description{
Add default values to render variables that won't get called every timestep
}
