% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{llin_impact}
\alias{llin_impact}
\title{LLIN impact}
\usage{
llin_impact(type, target, p, individuals, variables, api)
}
\arguments{
\item{type}{Infection type}

\item{target}{Individual indices}

\item{p}{Condition specific parameters}

\item{individuals}{Model individuals}

\item{variables}{Model variables}

\item{api}{Model api}
}
\description{
Estimate modifier due to LLIN impact
}
