% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{llin_impact}
\alias{llin_impact}
\title{LLIN impact}
\usage{
llin_impact(disease, target, p, variables)
}
\arguments{
\item{disease}{Infection disease}

\item{target}{Individual indices}

\item{p}{Condition specific parameters}

\item{variables}{Model variables}
}
\description{
Estimate modifier due to LLIN impact
}
