% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{long_symptoms}
\alias{long_symptoms}
\title{Long symptoms}
\usage{
long_symptoms(target, disease_index, threshold, timestep, variables)
}
\arguments{
\item{target}{Children seeking treatment}

\item{disease_index}{Disease index}

\item{threshold}{Period after which illness is classified as severe}

\item{timestep}{Timestep}

\item{variables}{Model variables}
}
\value{
Bitset
}
\description{
Returns a bitset of those who have had symptoms for a period > the threshold to define illness as severe
}
