% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{long_symptoms}
\alias{long_symptoms}
\title{Long symptoms}
\usage{
long_symptoms(target, disease_type, threshold, timestep, variables, parameters)
}
\arguments{
\item{target}{Children seeking treatment}

\item{threshold}{Period after which illness is classified as severe}

\item{timestep}{Timestyep}

\item{symptom_start_var}{Variable of symptom start time}
}
\value{
Bitset
}
\description{
Returns a bitset of those who have had symptoms for a period > the threshold to define illness as severe
}
