% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{recover_event}
\alias{recover_event}
\title{Recovery event}
\usage{
recover_event(variables, condition)
}
\arguments{
\item{variables}{Model variables}

\item{condition}{Condition recovered from}
}
\value{
Event
}
\description{
A recovery event. Resets status back to 0 (susceptible) and clears the disease index.
}
