% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{render_demography}
\alias{render_demography}
\title{Render demographic outputs}
\usage{
render_demography(variables, renderer)
}
\arguments{
\item{variables}{Model variables}

\item{renderer}{Model renderer}
}
\description{
Average ages and number of children in each year age-group.
}
