% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{render_incidence}
\alias{render_incidence}
\title{Record incidence}
\usage{
render_incidence(new_infections, condition, diseases, timestep, renderer)
}
\arguments{
\item{new_infections}{Indices of new infections}

\item{condition}{Condition: diarrhoea, malaria or pneumonia}

\item{diseases}{vector of diseases for condition}

\item{timestep}{Current time}

\item{renderer}{Model renderer}
}
\description{
Record overall (condition) and disaggregated (disease) incidence of new infection
}
