% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_prevalence}
\alias{render_prevalence}
\title{Record prevalence}
\usage{
render_prevalence(variables, renderer, parameters)
}
\description{
Record overall (condition) and disaggregated (disease) prevalence
}
