% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{replace_child}
\alias{replace_child}
\title{Replace child with new}
\usage{
replace_child(timestep, variables, target, parameters)
}
\arguments{
\item{timestep}{Current time}

\item{variables}{Model variables}

\item{target}{Target indices}

\item{parameters}{Model parameters}
}
\description{
Replace child with new
}
