% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{replace_child}
\alias{replace_child}
\title{Replace child with new}
\usage{
replace_child(api, target, individuals, variables)
}
\arguments{
\item{api}{Model API}

\item{target}{Indices of children to replace}

\item{individuals}{Model individuals}

\item{variables}{Model variables}
}
\description{
Replace child with new
}
