% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{rtexp}
\alias{rtexp}
\title{Random deviates from upper-truncated exponential distribution}
\usage{
rtexp(n = 1, rate = 1, upper = Inf)
}
\arguments{
\item{n}{Number of draws}

\item{rate}{Rate}

\item{upper}{Upper truncation (inclusive: (0, upper])}
}
\value{
Random draws from truncated exponential
}
\description{
Random deviates from upper-truncated exponential distribution
}
