% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{severe_diagnosis}
\alias{severe_diagnosis}
\title{Diagnostic result for severe disease}
\usage{
severe_diagnosis(target, sens, spec, parameters, variables, disease_index)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=diagnosis]{diagnosis}}
  \describe{
    \item{\code{target}}{Children visiting provider}
    \item{\code{sens}}{Diagnostic sensitivity}
    \item{\code{spec}}{Diagnostic specificity}
    \item{\code{disease_index}}{Index of diseases of target type}
  }}
}
\value{
A bitset of those testing positive
}
\description{
Diagnostic result for severe disease
}
