% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{severe_diagnosis}
\alias{severe_diagnosis}
\title{Diagnostic result}
\usage{
severe_diagnosis(target, disease_type, sens, spec, variables, parameters)
}
\arguments{
\item{target}{Children visiting provider}

\item{sens}{Diagnostic sensitivity}

\item{spec}{Diagnostic specificity}

\item{status_variable}{Disease status variable}

\item{positive}{Status categories that are true +ves}

\item{negative}{Status categories  that are true -ves}
}
\value{
Bitset
}
\description{
Returns a bitset of those testing positive
}
