% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{time_since_onset}
\alias{time_since_onset}
\title{Time since onset}
\usage{
time_since_onset(target, condition, variables, timestep)
}
\arguments{
\item{target}{Target children}

\item{condition}{Condition}

\item{variables}{Model variables}

\item{timestep}{Timestep}
}
\value{
Vector of times
}
\description{
Estimate the time since the onset of symptoms for a given condition
}
