% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{treatment_prophylaxis}
\alias{treatment_prophylaxis}
\title{Treatment prophylaxis modifier}
\usage{
treatment_prophylaxis(time_since_treatment, pr_hl)
}
\arguments{
\item{time_since_treatment}{Time since disease last treated}

\item{pr_hl}{Prophylaxis half life}
}
\value{
Treatment prophylaxis effect
}
\description{
Treatment prophylaxis modifier
}
