% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{type_index}
\alias{type_index}
\title{Index diseases of a given type}
\usage{
type_index(parameters, type)
}
\arguments{
\item{parameters}{Model parameters}

\item{type}{Disease type (diarrhoea, malaria or pneumonia)}
}
\value{
A vector indexing which disease are of requested type
}
\description{
Index diseases of a given type
}
