% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{vaccine_impact}
\alias{vaccine_impact}
\title{Vaccine impact}
\usage{
vaccine_impact(target, disease, ages, parameters, variables)
}
\arguments{
\item{ages}{Individuals ages}

\item{p}{Condition specific parameters}

\item{variable}{Vaccine coverage variable for specific disease}
}
\value{
Vaccine modifier
}
\description{
Estimate modifier due to vaccine impact
}
