% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{vaccine_impact}
\alias{vaccine_impact}
\title{Vaccine impact}
\usage{
vaccine_impact(type, target, ages, p, individuals, variables, api)
}
\arguments{
\item{type}{Infection type}

\item{target}{Individual indices}

\item{ages}{Individuals ages}

\item{p}{Condition specific parameters}

\item{individuals}{Model individuals}

\item{variables}{Model variables}

\item{api}{Model api}
}
\value{
Vaccine modifier
}
\description{
Estimate modifier due to vaccine impact
}
