% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{vaccine_impact}
\alias{vaccine_impact}
\title{Vaccine impact}
\usage{
vaccine_impact(disease, index, target, ages, p, variables)
}
\arguments{
\item{disease}{Infection disease}

\item{index}{Infection disease index}

\item{target}{Individual indices}

\item{ages}{Individuals ages}

\item{p}{Condition specific parameters}

\item{variables}{Model variables}
}
\value{
Vaccine modifier
}
\description{
Estimate modifier due to vaccine impact
}
